Add cagra search to vamana build graph tests

+++ cpp/test/neighbors/ann_vamana/test_int8_t_uint32_t.cu
@@ -20,7 +20,7 @@
 
 namespace cuvs::neighbors::vamana {
 
-typedef AnnVamanaTest<int8_t, float, std::uint32_t> AnnVamanaTestI8_U32;
+typedef AnnVamanaTest<float, int8_t, std::uint32_t> AnnVamanaTestI8_U32;
 TEST_P(AnnVamanaTestI8_U32, AnnVamana) { this->testVamana(); }
 
 INSTANTIATE_TEST_CASE_P(AnnVamanaTest, AnnVamanaTestI8_U32, ::testing::ValuesIn(inputs));
